Refactor SliceIterator.
