+++ tket/include/tket/Circuit/Circuit.hpp
@@ -46,6 +46,7 @@
 #include "Conditional.hpp"
 #include "DAGDefs.hpp"
 #include "ResourceData.hpp"
+#include "Slices.hpp"
 #include "tket/Gate/OpPtrFunctions.hpp"
 #include "tket/Utils/Json.hpp"
 #include "tket/Utils/SequencedContainers.hpp"
@@ -55,7 +56,6 @@ namespace tket {
 
 typedef std::vector<EdgeVec> BundleVec;
 
-typedef VertexVec Slice;
 typedef std::vector<Slice> SliceVec;
 
 typedef std::vector<VertPort> QPathDetailed;
@@ -107,11 +107,6 @@ typedef boost::multi_index::multi_index_container<
             boost::multi_index::const_mem_fun<
                 BoundaryElement, std::string, &BoundaryElement::reg_name>>>>
     boundary_t;
-
-typedef sequenced_map_t<UnitID, Edge> unit_frontier_t;
-// typedef sequenced_map_t<Qubit, Edge> q_frontier_t;
-typedef sequenced_map_t<Bit, EdgeVec> b_frontier_t;
-// typedef sequenced_map_t<Bit, Edge> c_frontier_t;
 
 typedef std::unordered_map<unsigned, unsigned> permutation_t;
 
@@ -171,17 +166,6 @@ struct TraversalPoint {
   Vertex from;
   Vertex to;
   Edge edge;
-};
-
-struct CutFrontier {
-  std::shared_ptr<Slice> slice;
-  std::shared_ptr<unit_frontier_t> u_frontier;
-  std::shared_ptr<b_frontier_t> b_frontier;
-  void init() {
-    slice = std::make_shared<Slice>();
-    u_frontier = std::make_shared<unit_frontier_t>();
-    b_frontier = std::make_shared<b_frontier_t>();
-  }
 };
 
 // list of error types to throw out
@@ -253,53 +237,6 @@ class Circuit {
       E_iterator &eend) const;
 
  public:
-  /*SliceIterator class is used for lazy evaluation of slices */
-  class SliceIterator {
-   public:  // these are currently public to allow skip_func slicing.
-    CutFrontier cut_;
-    std::shared_ptr<b_frontier_t> prev_b_frontier_;
-    const Circuit *circ_;
-
-    class Sliceholder {
-     private:
-      Slice current_slice_;
-
-     public:
-      explicit Sliceholder(Slice slice) : current_slice_(slice) {}
-      Slice operator*() const { return current_slice_; }
-    };
-
-    // take in an unsigned 'n' and a circuit and give the 'n'th slice
-    // note: n=0 gives an empty SliceIterator
-    // n=1 gives the first slice
-
-    SliceIterator(
-        const Circuit &circ, const std::function<bool(Op_ptr)> &skip_func);
-    explicit SliceIterator(const Circuit &circ);
-    SliceIterator() : cut_(), circ_() { cut_.init(); }
-    Slice operator*() const { return *cut_.slice; }
-    bool operator==(const SliceIterator &other) const {
-      return *cut_.slice == *other.cut_.slice;
-    }
-    bool operator!=(const SliceIterator &other) const {
-      return !(*this == other);
-    }
-    std::shared_ptr<const unit_frontier_t> get_u_frontier() const {
-      return cut_.u_frontier;
-    }
-    std::shared_ptr<const b_frontier_t> get_b_frontier() const {
-      return cut_.b_frontier;
-    }
-    std::shared_ptr<const b_frontier_t> get_prev_b_frontier() const {
-      return prev_b_frontier_;
-    }
-    // A postfix increment operator overload
-    Sliceholder operator++(int);
-    // A prefix increment operator overload
-    SliceIterator &operator++();
-    bool finished() const;
-  };
-
   SliceIterator slice_begin() const;
   static SliceIterator slice_end();
   static const SliceIterator nullsit;
